PS-7309 : Data Masking Plugin: gen_range() doesn't support negative numbers
Problem:
--------
gen_range() when given a range with negative numbers, it produces a
positive number

Example: SELECT gen_range(-1000, -800); -> 4294966357

Fix:
----
1. Support negative numbers.
2. Support a range from negative and postive. ie gen_range(-100, 100) is valid
3. We now support a range of long. So the valid range is -9223372036854775808 to
   9223372036854775807

+++ plugin/data_masking/src/udf/udf_utils.cc
@@ -70,10 +70,10 @@ std::string random_number(const unsigned int length) {
   return str;
 }
 
-unsigned int random_number(const unsigned int min, const unsigned int max) {
+long random_number(const long min, const long max) {
   std::random_device r;
   std::default_random_engine el(r());
-  std::uniform_int_distribution<unsigned int> dist(min, max);
+  std::uniform_int_distribution<long> dist(min, max);
 
   return dist(el);
 }
